Add auth guard for CRUD operation in expense

+++ src/expense/expense.controller.ts
@@ -8,31 +8,35 @@ import {
   Delete,
   NotFoundException,
   ParseUUIDPipe,
+  UseGuards,
+  ForbiddenException,
+  HttpCode,
 } from "@nestjs/common";
 import { ExpenseService } from "./expense.service";
 import { CreateExpenseDto } from "./dto/create-expense.dto";
 import { UpdateExpenseDto } from "./dto/update-expense.dto";
-import * as dayjs from "dayjs";
+import { CurrentUser } from "src/auth/current-user.decorator";
+import { User } from "src/user/entities/user.entity";
+import { JwtAuthGuard } from "src/auth/jwt-auth.guard";
+import { Expense } from "./entities/expense.entity";
 
 @Controller("expense")
 export class ExpenseController {
   constructor(private readonly expenseService: ExpenseService) {}
 
   @Post()
-  async create(@Body() createExpenseDto: CreateExpenseDto) {
-    const formattedDate = dayjs(createExpenseDto.date).format("YYYY-MM-DD");
-    const validatedExpense = { ...createExpenseDto, date: formattedDate };
-    return await this.expenseService.create(validatedExpense);
+  @UseGuards(JwtAuthGuard)
+  async create(
+    @Body() createExpenseDto: CreateExpenseDto,
+    @CurrentUser() user: User,
+  ) {
+    return await this.expenseService.create(createExpenseDto, user);
   }
 
   @Get()
-  async findAll() {
-    return await this.expenseService.findAll();
-  }
-
-  @Get("ytd")
-  findYtd() {
-    return this.expenseService.findExpensesBeforeToday();
+  @UseGuards(JwtAuthGuard)
+  async findAll(@CurrentUser() user: User) {
+    return await this.expenseService.findAll(user);
   }
 
   @Get(":id")
@@ -41,9 +45,11 @@ export class ExpenseController {
   }
 
   @Patch(":id")
+  @UseGuards(JwtAuthGuard)
   async update(
     @Param("id", ParseUUIDPipe) id: string,
     @Body() updateExpenseDto: UpdateExpenseDto,
+    @CurrentUser() user: User,
   ) {
     const expense = await this.expenseService.findOne(id);
 
@@ -51,11 +57,49 @@ export class ExpenseController {
       throw new NotFoundException();
     }
 
+    if (expense.userId !== user.userId) {
+      throw new ForbiddenException(
+        null,
+        "You are not authorized to change this expense",
+      );
+    }
+
     return this.expenseService.update(id, updateExpenseDto);
   }
 
   @Delete(":id")
-  remove(@Param("id", ParseUUIDPipe) id: string) {
+  @UseGuards(JwtAuthGuard)
+  @HttpCode(204)
+  async remove(
+    @Param("id", ParseUUIDPipe) id: string,
+    @CurrentUser() user: User,
+  ) {
+    const expense = await this.expenseService.findOne(id);
+
+    if (!expense) {
+      throw new NotFoundException();
+    }
+
+    if (expense.userId !== user.userId) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to remove this expense`,
+      );
+    }
+
     return this.expenseService.remove(id);
+  }
+
+  // no need this d
+  @Get("user/:userId")
+  async findAllExpensesByUserId(
+    @Param("userId") userId: string,
+  ): Promise<Expense[]> {
+    return this.expenseService.findAllExpensesByUserId(userId);
+  }
+  // trying out query builder
+  @Get("ytd")
+  findYtd() {
+    return this.expenseService.findExpensesBeforeToday();
   }
 }
